% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_diameter_darcy.R
\name{calc_diameter_darcy}
\alias{calc_diameter_darcy}
\title{Calculate Required Pipe Diameter using Darcy-Weisbach}
\usage{
calc_diameter_darcy(
  loss,
  length,
  flow,
  roughness,
  friction_fun = calc_friction_cw,
  viscosity = 1.004e-06,
  gravity = 9.81
)
}
\arguments{
\item{loss}{Numeric. The target friction head loss (meters).}

\item{length}{Numeric. The length of the pipe (meters).}

\item{flow}{Numeric. The volumetric flow rate (cubic meters per second).}

\item{roughness}{Numeric. The absolute internal roughness of the pipe (meters). Required unless \code{friction_factor} is provided.}

\item{friction_fun}{Function. A function to calculate the friction factor. It must accept \code{reynolds}, \code{roughness}, and \code{diameter} as arguments. Defaults to \code{calc_friction_cw}.}

\item{viscosity}{Numeric. Kinematic viscosity of the fluid (sq. meters per sec). Default is 1.004e-6.}

\item{gravity}{Numeric. Acceleration due to gravity (meters per second squared). Default is 9.81.}
}
\value{
A numeric vector representing the required internal diameter in meters.
}
\description{
Calculates the required internal pipe diameter iteratively given a
specific head loss and flow rate. By taking a functional approach, it supports
any friction factor calculation function.
}
\examples{
# Find diameter using the default Colebrook-White function
calc_diameter_darcy(loss = 5, length = 100, flow = 0.02, roughness = 0.00026)

# Find diameter by injecting Swamee-Jain
calc_diameter_darcy(
  loss = 5, length = 100, flow = 0.02, roughness = 0.00026,
  friction_fun = calc_friction_sj
)

}
