% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_flow_darcy.R
\name{calc_flow_darcy}
\alias{calc_flow_darcy}
\title{Calculate Flow Rate using Darcy-Weisbach}
\usage{
calc_flow_darcy(
  loss,
  length,
  diameter,
  roughness,
  friction_fun = calc_friction_cw,
  viscosity = 1.004e-06,
  gravity = 9.81
)
}
\arguments{
\item{loss}{Numeric. The target friction head loss (meters).}

\item{length}{Numeric. The length of the pipe (meters).}

\item{diameter}{Numeric. The internal diameter of the pipe (meters).}

\item{roughness}{Numeric. The absolute internal roughness of the pipe (meters). Required unless \code{friction_factor} is provided.}

\item{friction_fun}{Function. A function to calculate the friction factor. It must accept \code{reynolds}, \code{roughness}, and \code{diameter} as arguments. Defaults to \code{calc_friction_cw}.}

\item{viscosity}{Numeric. Kinematic viscosity of the fluid (sq. meters per sec). Default is 1.004e-6.}

\item{gravity}{Numeric. Acceleration due to gravity (meters per second squared). Default is 9.81.}
}
\value{
A numeric vector representing the volumetric flow rate in cubic meters per second.
}
\description{
Calculates the volumetric flow rate iteratively given a specific
head loss. This function uses a functional programming approach, allowing the
injection of any friction factor function (e.g., Colebrook-White).
}
\examples{
# Default: Uses Colebrook-White to find the flow rate
calc_flow_darcy(loss = 5, length = 100, diameter = 0.1, roughness = 0.00026)

# Injecting Swamee-Jain function
calc_flow_darcy(
  loss = 5, length = 100, diameter = 0.1, roughness = 0.00026,
  friction_fun = calc_friction_sj
)

}
