% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_friction_sj.R
\name{calc_friction_sj}
\alias{calc_friction_sj}
\title{Calculate Friction Factor using Swamee-Jain}
\usage{
calc_friction_sj(reynolds, roughness, diameter)
}
\arguments{
\item{reynolds}{Numeric. The Reynolds number of the flow (dimensionless).}

\item{roughness}{Numeric. The absolute internal roughness of the pipe (meters).}

\item{diameter}{Numeric. The internal diameter of the pipe (meters).}
}
\value{
A numeric vector representing the Darcy friction factor.
}
\description{
Calculates the Darcy friction factor using the explicit Swamee-Jain
equation. This is a highly accurate approximation of the Colebrook-White equation
that does not require iteration.
}
\examples{
calc_friction_sj(reynolds = 100000, roughness = 0.00026, diameter = 0.1)

}
