% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diameter.R
\name{Diameter}
\alias{Diameter}
\title{Diameter of pipes}
\usage{
Diameter(J, L = 1, Q, RC, Eq = "CW", friend = FALSE, v = 1.01e-06,
  g = 9.81, x1 = 0.0625)
}
\arguments{
\item{J}{Unitary head loss  in meters per meter or Total head loss in meters}

\item{L}{Length of pipe in meters. If \code{J} is unitary head loss this parameter
is unnecessary.}

\item{Q}{Flow rate in cubic meters per second}

\item{RC}{Roughness coefficient. Absolute roughness (E in meters) for
Darcy-Weisbach equation or C (dimensionless) for Hazen-Willians equation or
b (dimensionless) for Flamant equation}

\item{Eq}{To choose the equation that will be used. By default \code{Eq="CW"} for
Colebrook-White. Other options are:  \code{Eq="SJ"} for Swamee-Jain; \code{Eq="HW"} for
Hazen-Willians; \code{Eq="FL"} for Flamant}

\item{friend}{If friend=TRUE the results are presented in a friendly way}

\item{v}{Kinematic viscosity of fluiyd in square meters per second. By default use the
value for water at 20oC \code{v=1.01e-6 m2/s}.  Unnecessary for empirical equations.}

\item{g}{Gravitational acceleration g=9.81 meters per square second.  Unnecessary for
empirical equations.}

\item{x1}{Initial parameter of f for Newthon-Raphson. By default \code{x1=0.0625}.
Unnecessary for empirical or explicity equations.}
}
\value{
D Diameter of pipe in meters
}
\description{
\code{Diameter} calculates the diameter of pipes for a given head loss
}
\examples{
Diameter(J=0.197,Q=0.005,RC=1e-4,Eq="CW", friend=TRUE)
Diameter(J=0.197,Q=0.005,RC=1e-4,Eq="CW")
Diameter(J=0.197,Q=0.005,RC=1e-4,Eq="SJ", friend=TRUE)
Diameter(J=0.197,Q=0.005,RC=140, Eq="HW", friend=TRUE)
Diameter(J=0.197,Q=0.005,RC=140, Eq="HW")
}

