% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_rate.R
\name{flow_rate}
\alias{flow_rate}
\title{Flow rate}
\usage{
flow_rate(hf, D, L, RC, v = 1.01e-06, g = 9.81)
}
\arguments{
\item{hf}{Total head loss in meters}

\item{D}{Diameter in meters}

\item{L}{Length of pipe in meters}

\item{RC}{Roughness coefficient. Absolute roughness (in meters) for
Universal equation.}

\item{v}{Kinematic viscosity of fluid in square meters per second. By
default use the value for water at 20 Celsius degree
(\code{v = 1.01e-6 m^2/s}). Unnecessary for empirical
equations.}

\item{g}{Gravitational acceleration. By default use the value
\code{g = 9.81 m/s^2}}
}
\value{
Q Flow rate in cubic meters per second
}
\description{
\code{flow_rate} calculate the flow rate for a given head loss
}
\examples{
flow_rate(hf = 0.97, L = 10, D = 0.008, v = 3e-6)
flow_rate(hf = 9.7, L = 10, D = 0.050, RC = 0.0001)
}
