% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowRate.R
\name{flowRate}
\alias{flowRate}
\title{Flow rate}
\usage{
flowRate(J, L = 1, D, RC, Vel = FALSE, Eq = "CW", friend = FALSE,
  v = 1.01e-06, g = 9.81)
}
\arguments{
\item{J}{Unitary head loss  in meters per meter or Total head loss in meters}

\item{L}{Length of pipe in meters. If \code{J} is unitary head loss this parameter
is unnecessary.}

\item{D}{Diamter of pipe in meters}

\item{RC}{Roughness coefficient. Absolute roughness (E in meters) for
Darcy-Weisbach equation or C (dimensionless) for Hazen-Willians equation or
b (dimensionless) for Flamant equation}

\item{Vel}{Velocity. If \code{Vel=TRUE} then the return are the flow velocity.}

\item{Eq}{To choose the equation that will be used. By default \code{Eq="CW"} for
Colebrook-White. Other options are:  \code{Eq="SJ"} for Swamee-Jain; \code{Eq="HW"} for
Hazen-Willians; \code{Eq="FL"} for Flamant}

\item{friend}{If \code{friend=TRUE} the results are presented in a friendly way}

\item{v}{Kinematic viscosity of fluiyd in square meters per second. By default use the
value for water at 20oC \code{v=1.01e-6 m2/s}.  Unnecessary for empirical equations.}

\item{g}{Gravitational acceleration g=9.81 meters per square second.  Unnecessary for
empirical equations.}
}
\value{
Q Flow rate in cubic meters per second or V velocity flow in meters per second
}
\description{
\code{flowRate} calculate the flow rate or flow velocity for a given head loss
}
\examples{
flowRate(J=0.097,D=0.005,RC=1e-4, Vel=TRUE, Eq="CW", friend=TRUE)
flowRate(J=0.097,D=0.005,RC=1e-4, Eq="CW", friend=TRUE)
flowRate(J=0.097,D=0.05,RC=1e-4, Eq="SJ")
flowRate(J=0.097,D=0.005,RC=140, Eq="HW")
}

