% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_loss.R
\name{head_loss}
\alias{head_loss}
\title{Head loss over the length of pipe}
\usage{
head_loss(D, Q, L, RC, v = 1.01e-06, g = 9.81, x1 = 0.06)
}
\arguments{
\item{D}{Diameter in meters}

\item{Q}{Flow rate in cubic meters per second}

\item{L}{Length of pipe in meters}

\item{RC}{Roughness coefficient. Absolute roughness (in meters) for
Universal equation.}

\item{v}{Kinematic viscosity of fluid in square meters per second. By
default use the value for water at 20 Celsius degree
(\code{v = 1.01e-6 m^2/s}). Unnecessary for empirical
equations.}

\item{g}{Gravitational acceleration. By default use the value
\code{g = 9.81 m/s^2}}

\item{x1}{Initial parameter of f for Newton-Raphson.
By default \code{x1=0.06}.
Unnecessary for empirical or explicit equations.}
}
\value{
hf Head loss in meters
}
\description{
\code{headLoss} calculate the head loss by friction in pipes by different
methods
}
\examples{
head_loss(D = 0.025, Q = 0.000000001, L = 200)
head_loss(D = 0.050, Q = 0.0006, L = 200, RC = 0.0001)
}
