% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headLoss.R
\name{headLoss}
\alias{headLoss}
\title{Head loss in meters over the length of pipe}
\usage{
headLoss(D, Q, L, RC, Eq = "CW", friend = FALSE, v = 1.01e-06, g = 9.81,
  x1 = 0.0625)
}
\arguments{
\item{D}{Diameter in meters}

\item{Q}{Flow rate in cubic meters per second}

\item{L}{Length of pipe in meters}

\item{RC}{Roughness coefficient. Absolute roughness (E in meters) for
Darcy-Weisbach equation or C (dimensionless) for Hazen-Willians equation or
b (dimensionless) for Flamant equation}

\item{Eq}{To choose the equation that will be used. By default \code{Eq="CW"} for
Colebrook-White. Other options are:  \code{Eq="SJ"} for Swamee-Jain; \code{Eq="HW"} for
Hazen-Willians; \code{Eq="FL"} for Flamant}

\item{friend}{If friend=TRUE the results are presented in a friendly way}

\item{v}{Kinematic viscosity of fluiyd in square meters per second. By
default use the value for water at 20ºC \code{v=1.01e-6 m2/s}.  Unnecessary for empirical equations.}

\item{g}{Gravitational acceleration \code{g=9.81} meters per square second.  Unnecessary for empirical equations.}

\item{x1}{Initial parameter of f for Newthon-Raphson. By default \code{x1=0.0625}.  Unnecessary for empirical or explicity equations.}
}
\value{
Hf Head loss in meters
}
\description{
\code{headLoss} calculate the head loss by fricition in pipes by different
methods
}
\examples{
headLoss(D=0.025,Q=0.000000001,L=200,RC=0.0001,Eq="CW",friend=TRUE)
headLoss(D=0.025,Q=0.000000001,L=200,RC=140,Eq="HW",friend=TRUE)
headLoss(D=0.025,Q=0.001,L=200,RC=0.0001,Eq="CW",friend=TRUE)
headLoss(D=0.025,Q=0.001,L=200,RC=0.0001,Eq="SJ",friend=TRUE)
headLoss(D=0.200,Q=0.001,L=3200,RC=130,Eq="HW",friend=TRUE)
headLoss(D=0.025,Q=0.001,L=200,RC=0.000135,Eq="FL",friend=TRUE)

}

