% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re.R
\name{re}
\alias{re}
\title{Reynolds Number}
\usage{
re(D, Q, v)
}
\arguments{
\item{D}{Diameter in meters}

\item{Q}{Flow rate in cubic meters per second}

\item{v}{Kinematic viscosity of fluid in square meters per second.}
}
\value{
re Reynolds number
}
\description{
\code{re} calculates the Reynolds number
}
\examples{
re(D = 0.025, Q = 0.000000001, v = 1.01e-6)
re(D = 0.050, Q = 0.0006, v = 1.01e-6)
}
