% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diameter.R
\name{diameter}
\alias{diameter}
\title{Diameter of pipes}
\usage{
diameter(hf, Q, L, RC, v = 1.01e-06, g = 9.81, x1 = 0.06)
}
\arguments{
\item{hf}{Total head loss in meters}

\item{Q}{Flow rate in cubic meters per second}

\item{L}{Length of pipe in meters}

\item{RC}{Roughness coefficient. Absolute roughness (in meters) for
Universal equation.}

\item{v}{Kinematic viscosity of fluid in square meters per second. By
default use the value for water at 20 Celsius degree
(\code{v = 1.01e-6 m^2/s}). Unnecessary for empirical
equations.}

\item{g}{Gravitational acceleration. By default use the value
\code{g = 9.81 m/s^2}}

\item{x1}{Initial parameter of f for Newton-Raphson.
By default \code{x1=0.06}.
Unnecessary for empirical or explicit equations.}
}
\value{
D Diameter of pipe in meters
}
\description{
\code{diameter} calculates the diameter of pipes for a given head loss
}
\examples{
diameter(hf = 19.7, L = 100, Q = 0.0000005)
diameter(hf = 9.7, L = 10, Q = 0.005, RC = 0.0001)
}
